Create DB Schema to share with server - TS definitions
