+++ client/src/pages/AllItems.tsx
@@ -4,7 +4,8 @@ import { useFind } from 'use-pouchdb';
 import { useRef } from 'react';
 import { useLists, useItems } from '../components/Usehooks';
 import SyncIndicator from '../components/SyncIndicator';
-import { HistoryProps, ItemDoc, ListCombinedRow, ListRow, RowType } from '../components/DataTypes';
+import { HistoryProps} from '../components/DataTypes';
+import { ItemDoc } from '../components/DBSchema';
 import './AllItems.css';
 import ErrorPage from './ErrorPage';
