+++ client/src/pages/GlobalItem.tsx
@@ -6,7 +6,8 @@ import { useFind, usePouch } from 'use-pouchdb';
 import { useState,  useContext, useRef } from 'react';
 import { useGetOneDoc } from '../components/Usehooks';
 import './Category.css';
-import {  HistoryProps,  CategoryDoc, UomDoc } from '../components/DataTypes';
+import {  HistoryProps} from '../components/DataTypes';
+import { CategoryDoc, UomDoc } from '../components/DBSchema';
 import SyncIndicator from '../components/SyncIndicator';
 import { closeOutline} from 'ionicons/icons';
 import ErrorPage from './ErrorPage';
